move hungarian rings
