+++ cayleypy/graphs_lib.py
@@ -6,7 +6,7 @@ from itertools import permutations
 from warnings import warn
 
 from .cayley_graph_def import CayleyGraphDef, MatrixGenerator
-from .hungarian_rings import hungarian_rings_generators
+from cayleypy.puzzles.hungarian_rings import hungarian_rings_generators
 from .permutation_utils import (
     transposition,
     permutation_from_cycles as pfc,
@@ -329,26 +329,19 @@ class PermutationGroups:
 
 
 def prepare_graph(name: str, n: int = 0) -> CayleyGraphDef:
-    """Returns pre-defined Cayley or Schreier coset graph.
+    """Returns pre-defined graph by codename.
+
+    This function is a wrapper around other functions that produce graphs.
 
     TODO: move everything out and remove this function.
     DO NOT ADD ANY NEW GRAPHS HERE!
 
-    Supported graphs:
-      * "cube_2/2/2_6gensQTM" - Schreier coset graph for 2x2x2 Rubik's cube with fixed back left upper corner and only
-          quarter-turns allowed. There are 6 generators (front, right, down face - clockwise and counterclockwise).
-      * "cube_2/2/2_9gensHTM" - same as above, but allowing half-turns (it has 9 generators).
-      * "cube_3/3/3_12gensQTM" - Schreier coset graph for 3x3x3 Rubik's cube with fixed central pieces and only
-          quarter-turns allowed. There are 12 generators (clockwise and counterclockwise rotation for each face).
-      * "cube_3/3/3_18gensHTM" - same as above, but allowing half-turns (it has 18 generators).
       * "mini_pyramorphix" – Cayley graph for a subgroup of S_24, acting on 24 titles. It is generated by 17 moves
           inspired by a simplified version of the Pyramorphix puzzle. Moves are based on overlapping 2- and 3-cycles
           and result in a symmetric, undirected graph. (order of the graph 24, degree 17, order of the group 288)
       * "pyraminx" - Cayley graph for a subgroup pf S_36 acting on 36 tiles. It is generated by 8 elements inspired
           by Pyraminx puzzle. 4 elemets represent rotations  of the tetrahedron tips, while 4 others -- rotations
           of its base layers.
-      * "hungarian_rings" - Cayley graph for S_n (n>=4), generated by rotating two rings in both directions.
-          For each ring structure and their intersection it has four generators.
       * "starminx" - Cayley graph, generated by 20 moves of Starminx puzzle. Each move is a rotation of a
           corner, centered around one of the dodecahedron’s pentagonal faces (face centres never move). A single turn
           corresponds to a set of disjoint 3-cycles, each on one triangular sticker.
@@ -379,11 +372,7 @@ def prepare_graph(name: str, n: int = 0) -> CayleyGraphDef:
         central_state = list(range(len(generators[0])))
         return CayleyGraphDef.create(generators, central_state=central_state, generator_names=generator_names)
     elif name == "hungarian_rings":
-        assert n % 2 == 0
-        ring_size = (n + 2) // 2
-        assert ring_size >= 4
-        generators, generator_names = hungarian_rings_generators(ring_size=ring_size)
-        return CayleyGraphDef.create(generators, central_state=list(range(n)), generator_names=generator_names)
+        return Puzzles.hungarian_rings(n)
     elif name == "starminx":
         generator_names = list(STARMINX_MOVES.keys())
         generators = [STARMINX_MOVES[k] for k in generator_names]
