+++ cayleypy/puzzles/hungarian_rings_test.py
@@ -4,7 +4,7 @@ import os
 import pytest
 
 from .hungarian_rings import hungarian_rings_permutations, _circular_shift, _create_right_ring, _get_intersections
-from .permutation_utils import compose_permutations
+from cayleypy.permutation_utils import compose_permutations
 
 FAST_RUN = os.getenv("FAST") == "1"
