+++ cayleypy/puzzles/puzzles.py
@@ -3,6 +3,7 @@ from ..cayley_graph_def import CayleyGraphDef
 from ..permutation_utils import inverse_permutation, compose_permutations
 from .cube import rubik_cube_qstm, rubik_cube_htm, rubik_cube_qtm
 from .globe import globe_puzzle
+from .hungarian_rings import hungarian_rings_generators
 
 
 class Puzzles:
@@ -31,3 +32,15 @@ class Puzzles:
     def globe_puzzle(a: int, b: int) -> CayleyGraphDef:
         """Cayley graph for Globe puzzle group, a + 1 cycle and 2b order 2 generators."""
         return globe_puzzle(a, b)
+
+    @staticmethod
+    def hungarian_rings(n: int):
+        """Cayley graph for S_n (n>=4), generated by rotating two rings in both directions.
+
+        For each ring structure and their intersection it has four generators.
+        """
+        assert n % 2 == 0
+        ring_size = (n + 2) // 2
+        assert ring_size >= 4
+        generators, generator_names = hungarian_rings_generators(ring_size=ring_size)
+        return CayleyGraphDef.create(generators, central_state=list(range(n)), generator_names=generator_names)
